+++ src/pages/Home.tsx
@@ -18,7 +18,7 @@ export default function Home() {
     },
     onError: error => console.error('Login Failed:', error),
     scope:
-      'https://www.googleapis.com/auth/drive.file https://www.googleapis.com/auth/spreadsheets https://www.googleapis.com/auth/userinfo.profile https://www.googleapis.com/auth/userinfo.email',
+      'https://www.googleapis.com/auth/drive.file https://www.googleapis.com/auth/userinfo.profile https://www.googleapis.com/auth/userinfo.email',
   });
 
   return (